Move reloadrangethread into its own files
