+++ include/downloadthread.h
@@ -18,21 +18,6 @@ private:
 	std::vector<int> indexes;
 };
 
-class reloadrangethread {
-public:
-	reloadrangethread(Reloader* r,
-		unsigned int start,
-		unsigned int end,
-		unsigned int size,
-		bool unattended);
-	void operator()();
-
-private:
-	Reloader* reloader;
-	unsigned int s, e, ss;
-	bool u;
-};
-
 } // namespace newsboat
 
 #endif /* NEWSBOAT_DOWNLOADTHREAD_H_ */
