+++ include/reloadrangethread.h
@@ -1,0 +1,27 @@
+#ifndef NEWSBOAT_RELOADRANGETHREAD_H_
+#define NEWSBOAT_RELOADRANGETHREAD_H_
+
+#include <thread>
+
+#include "reloader.h"
+
+namespace newsboat {
+
+class reloadrangethread {
+public:
+	reloadrangethread(Reloader* r,
+		unsigned int start,
+		unsigned int end,
+		unsigned int size,
+		bool unattended);
+	void operator()();
+
+private:
+	Reloader* reloader;
+	unsigned int s, e, ss;
+	bool u;
+};
+
+} // namespace newsboat
+
+#endif /* NEWSBOAT_RELOADRANGETHREAD_H_ */
