+++ mk/mk.deps
@@ -77,16 +77,16 @@ src/controller.o: src/controller.cpp include/controller.h include/cache.h \
  include/filtercontainer.h include/fslock.h include/regexmanager.h \
  include/reloader.h include/remote_api.h include/urlreader.h \
  include/cliargsparser.h include/colormanager.h include/configcontainer.h \
- include/configparser.h include/downloadthread.h include/controller.h \
- include/exception.h include/exceptions.h include/feedhq_api.h \
- include/formatstring.h include/globals.h include/inoreader_api.h \
- include/logger.h include/newsblur_api.h rss/rsspp.h include/remote_api.h \
+ include/configparser.h include/downloadthread.h include/exception.h \
+ include/exceptions.h include/feedhq_api.h include/formatstring.h \
+ include/globals.h include/inoreader_api.h include/logger.h \
+ include/newsblur_api.h rss/rsspp.h include/remote_api.h \
  include/ocnews_api.h include/oldreader_api.h include/regexmanager.h \
  include/rss_parser.h include/stflpp.h include/strprintf.h \
  include/ttrss_api.h 3rd-party/json.hpp include/utils.h include/view.h \
- include/filebrowser_formaction.h include/formaction.h include/history.h \
- include/keymap.h include/stflpp.h include/htmlrenderer.h \
- include/textformatter.h xlicense.h
+ include/controller.h include/filebrowser_formaction.h \
+ include/formaction.h include/history.h include/keymap.h include/stflpp.h \
+ include/htmlrenderer.h include/textformatter.h xlicense.h
 src/dialogs_formaction.o: src/dialogs_formaction.cpp \
  include/dialogs_formaction.h include/formaction.h include/history.h \
  include/keymap.h include/configparser.h include/rss.h \
@@ -105,14 +105,7 @@ src/download.o: src/download.cpp include/download.h config.h \
  include/download.h include/fslock.h include/queueloader.h \
  include/pb_controller.h
 src/downloadthread.o: src/downloadthread.cpp include/downloadthread.h \
- include/controller.h include/cache.h include/configcontainer.h \
- include/configparser.h include/rss.h include/matcher.h \
- filter/FilterParser.h include/utils.h include/logger.h config.h \
- include/strprintf.h include/colormanager.h include/configpaths.h \
- include/cliargsparser.h include/feedcontainer.h \
- include/filtercontainer.h include/fslock.h include/regexmanager.h \
- include/reloader.h include/remote_api.h include/urlreader.h \
- include/logger.h
+ include/reloader.h include/logger.h config.h include/strprintf.h
 src/exception.o: src/exception.cpp include/exception.h config.h \
  include/exceptions.h include/configparser.h include/strprintf.h \
  include/utils.h include/configcontainer.h include/logger.h \
@@ -328,12 +321,14 @@ src/reloader.o: src/reloader.cpp include/reloader.h include/controller.h \
  include/configpaths.h include/cliargsparser.h include/feedcontainer.h \
  include/filtercontainer.h include/fslock.h include/regexmanager.h \
  include/reloader.h include/remote_api.h include/urlreader.h \
- include/downloadthread.h include/controller.h include/exceptions.h \
- include/reloadthread.h rss/rsspp.h include/remote_api.h \
- include/rss_parser.h rss/rsspp.h include/utils.h include/view.h \
- include/filebrowser_formaction.h include/formaction.h include/history.h \
- include/keymap.h include/stflpp.h include/htmlrenderer.h \
- include/textformatter.h include/formatstring.h
+ include/downloadthread.h include/exceptions.h include/formatstring.h \
+ include/reloadthread.h include/controller.h include/reloadrangethread.h \
+ rss/rsspp.h include/remote_api.h include/rss_parser.h rss/rsspp.h \
+ include/utils.h include/view.h include/filebrowser_formaction.h \
+ include/formaction.h include/history.h include/keymap.h include/stflpp.h \
+ include/htmlrenderer.h include/textformatter.h
+src/reloadrangethread.o: src/reloadrangethread.cpp \
+ include/reloadrangethread.h include/reloader.h
 src/reloadthread.o: src/reloadthread.cpp include/reloadthread.h \
  include/configcontainer.h include/configparser.h include/controller.h \
  include/cache.h include/rss.h include/matcher.h filter/FilterParser.h \
